ai_ops_assistant/requirements.txt
openai
requests
fastapi
uvicorn
python-dotenv
speechrecognition
pyttsx3
rich
cachetools
tiktoken
pydantic
